move forbidden, add helmets
