+++ src/pages/member/UserHome.js
@@ -1,8 +1,12 @@
 import React from 'react'
+import { Helmet } from 'react-helmet';
 
 function UserHome() {
     return (
         <div>
+            <Helmet>
+                <title>Biztech User Dashboard</title>
+            </Helmet>
             User Home
         </div>
     )
